+++ docs/changelog.rst
@@ -33,8 +33,9 @@ Changelog
 * :bug:`806` Force strings given to ``getpass`` during password prompts to be
   ASCII, to prevent issues on some platforms when Unicode is encountered.
   Thanks to Alex Louden for the patch.
-* :feature:`805` Update `~fabric.context_managers.shell_env` to play nice with
-  Windows (7, at least) systems. Thanks to Fernando Macedo for the patch.
+* :bug:`805` Update `~fabric.context_managers.shell_env` to play nice with
+  Windows (7, at least) systems and `~fabric.operations.local`. Thanks to
+  Fernando Macedo for the patch.
 * :bug:`654` Parallel runs whose sum total of returned data was large (e.g.
   large return values from the task, or simply a large number of hosts in the
   host list) were causing frustrating hangs. This has been fixed.